refactor: relocate agent utils
