+++ src/agent/modelHandlers/modelHandlerDeepSeek.ts
@@ -9,7 +9,7 @@ import { ModelHandlerOpenAI } from './modelHandlerOpenAI';
 import { ToolState } from '../core/ToolState';
 
 // Local imports - utilities
-import { convertContentToString } from '@utils/text/messageUtils';
+import { convertContentToString } from '@agent/utils/messageUtils';
 import { K_SLICE, MESSAGE_PREVIEW_LENGTH } from '@utils/config';
 
 // TODO: prompt_cache_hit_tokens can also be used here to correct the price and response usage computation in the base class (just overwrite the computePrice and computeResponseUsage methods with a revalues responseUsage.prompt_tokens_details?.cached_tokens and then call the super methods)
