+++ src/agent/modelHandlers/modelHandlerGoogleGenAI.ts
@@ -35,7 +35,7 @@ import { AgentLogger } from '@logger/AgentLogger';
 import { WorkspaceFS, AbsoluteFS } from '@utils/files';
 import { formatProviderError } from '@utils/sdkErrorUtils';
 import xmlUtils from '@utils/text/xmlUtils';
-import { calculateTokenPrice } from '@utils/priceUtils';
+import { calculateTokenPrice } from '@agent/utils/priceUtils';
 import { MESSAGE_TYPES } from '@logger/messageTypes';
 
 import { cleanFileContent } from '@replacement/engine';
