+++ src/agent/output/index.ts
@@ -1,4 +1,4 @@
 export * from './OutputHandler';
 export * from './IOutputHandler';
 export { NamedOutputFile } from './types';
-export { getOutputFileName } from '@utils/outputFileUtils';
+export { getOutputFileName } from '@agent/utils/outputFileUtils';
