+++ src/agent/utils/outputFileUtils.ts
