+++ src/agent/utils/priceUtils.ts
