+++ src/utils/text/index.ts
@@ -1,6 +1,4 @@
 export * from './stringUtils';
-export * from './messageUtils';
 export * from './xmlUtils';
 export * from './textEnhancementUtils';
-export * from './repetitionUtils';
 export { xmlUtils as default } from './xmlUtils';
